requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-yasg
drf-spectacular
drf-yasg